photos-processing/pyproject.toml
[tool.poetry]
name = "Ruurd Photos"
version = "2.0.0"
description = ""
authors = ["Ruurd Bijlsma <ruurd@bijlsma.dev>"]
license = "'"
readme = "README.md"
packages = [{ include = "*", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["all"], version = "^0.115.0" }
uvicorn = "^0.31.0"
httpx = "^0.27.2"
pillow = "^10.4.0"
watchdog = "^5.0.3"
tqdm = "^4.66.5"
pre-commit = "^3.8.0"
types-tqdm = "^4.66.0.20240417"
pydantic-settings = "^2.5.2"
pydantic = {extras = ["dotenv"], version = "^2.9.2"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"